Library query and upload. Deprecated python version
The Go based version of the library query and upload to the server side is now functional. A gzipped version will be sent to the server and a version will be saved locally as well.
