+++ client/config.json
@@ -1,0 +1,1 @@
+{"plex_token": "", "plex_server_url": "","server_url":""}
