--- example_config.json
@@ -1,6 +1,0 @@
-{
-    "plex_token": "af30987ag90ghf8s",
-    "plex_server_ip": "192.168.0.50",
-    "output_csv": "C:/Strom/ryanlibrarytest.csv",
-    "library_name": "Movies"
-}
