--- requirements.txt
@@ -1,2 +1,0 @@
-plexapi
-re
