+++ server/config.json
@@ -1,0 +1,1 @@
+{"port":"8080"}
